SCIP_Longint is a long long and must be mapped to jlong.
src/scipjni.i (SCIP_Longint): Fix the definition from "long int" to
"long long int", to match the actual definition in any current version
of SCIP (checked 7.0.3 and 8.0.0).

In particular, this is always a 64-bit integer, even on a 32-bit CPU,
and even on Windows (which always uses 32-bit long). So it also must be
mapped to the Java long (JNI jlong) type, not to Java int (JNI jint).
(SWIG defaults to mapping long to jint. This can be overridden, but that
would still not do the right thing on Windows, because the jlong would
still be truncated to a 32-bit C/C++ long before being passed to SCIP.)

src/scipjni_wrap.cxx, java/jscip/SCIPJNI.java,
java/jscip/SCIPJNIJNI.java: Regenerate.

java/jscip/Scip.java (setLongintParam): Change the type of the value
argument from int to long to match the above. Add a compatibility
overload with the old signature for binary compatibility.

+++ java/jscip/SCIPJNI.java
@@ -190,7 +190,7 @@ public class SCIPJNI {
     return SCIP_Retcode.swigToEnum(SCIPJNIJNI.SCIPsetIntParam(SWIGTYPE_p_SCIP.getCPtr(scip), name, value));
   }
 
-  public static SCIP_Retcode SCIPsetLongintParam(SWIGTYPE_p_SCIP scip, String name, int value) {
+  public static SCIP_Retcode SCIPsetLongintParam(SWIGTYPE_p_SCIP scip, String name, long value) {
     return SCIP_Retcode.swigToEnum(SCIPJNIJNI.SCIPsetLongintParam(SWIGTYPE_p_SCIP.getCPtr(scip), name, value));
   }
